+++ tests/dummy/app/JsonApi/Posts/Adapter.php
@@ -39,6 +39,13 @@ class Adapter extends AbstractAdapter
     ];
 
     /**
+     * @var array
+     */
+    protected $includePaths = [
+        'comments.created-by' => 'comments.user',
+    ];
+
+    /**
      * Adapter constructor.
      *
      * @param StandardStrategy $paging